Adjust the code and update Dockerfiles Put all the helpers into codes/common.go Use hub.opshub.sh/containerops/golang:1.8.3 as the base image
